services/sync/pyproject.toml
[project]
name = "jobl-sync"
version = "0.1.0"
description = "Job ingestion sync worker (scraper DB -> AI storage)"
requires-python = ">=3.11"
dependencies = [
  "sqlalchemy>=2.0.43",
  "pydantic-settings>=2.10.1",
  "psycopg[binary]>=3.2.9",
  "pymysql>=1.1.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
]

[project.scripts]
jobl-sync = "app.main:run"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
